Introduce release-tag profile
The release-tag profile is very minimal and exists to allow multi-module
projects to specify modules within the profile that might not be
specified at the project level. This is mainly applicable to complex
build setups where different operating systems are likely involved.

For the same reason, the maven-release-plugin does not perform any
verification steps; it simply runs a "clean". This is considered
reasonable given to be making a release it is expected the project
maintainer has run whatever tests or observed whatever CI environment is
appropriate to reach a release decision.