Release plugin to only run 'clean' phase
Given ubiquitous continuous integration use, there is marginal
incremental value in developers also running the verify phase during
`mvn -Prelease-tag release:clean release:prepare` operations. By
assuming a developer would be aware of the continuous integration
status of a project before preparing a release, removal of the verify
phase in this commit achieves useful goals:

1. Releases can be prepared much more quickly

2. Releases can be prepared without requiring the developer to install
   any native packages (eg Android NDK, cross-compilers etc)

3. Additional flexibility is offered in allowing sub-modules to be
   built by OS-specific images during CI (thus avoiding additional
   native packages such as mentioned above, plus any possible
   incompatibilities introduced through cross-compilation)

The disadvantage of this change is the risk a developer may prepare a
release for code that is incorrect. If this occurred the CI pipeline
would fail in its verify phase and as such the pipeline would fail
before any deployments are performed. A developer would simply need to
delete the tag and try again.