Run sortpom plugin before xml-format and avoid using test phase
Sortpom does not provide a configuration setting that offers short
elements such as <foo/>. Instead it reformats these to <foo />. The
XML formatter plugin does provide the preferred shorter format. It is
therefore necessary to run the XML formatter after the Sortpom plugin
so that its format is used. Prior to this commit sortpom would run in
the test phase and XML formatter in the prepare-package phase in order
to achieve this. However this meant the relatively common use of `mvn
test` would result in POMs that lack the desired formatting. This
commit therefore makes sortpom run in the prepare-package phase and
XML format run in the package phase, mean `mvn test` will not modify
the POM at all and `mvn package` will execute both plugins.